fixed bug in InterestControl
- condition was false as limit_ variable is unsigned int

+++ cpp/src/interest-control.cpp
@@ -80,7 +80,7 @@ InterestControl::decrement()
 bool
 InterestControl::increment()
 {
-	if (pipeline_ >= limit_)
+	if (pipeline_ >= (int)limit_)
 		return false;
 
 	pipeline_++;
